LPS-66929 Move Password Policy data handlers to the correct packages
