+++ modules/apps/foundation/password-policies-admin/password-policies-admin-impl/src/main/java/com/liferay/password/policies/admin/internal/exportimport/data/handler/PasswordPolicyStagedModelDataHandler.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.password.policies.admin.internal.lar;
+package com.liferay.password.policies.admin.internal.exportimport.data.handler;
 
 import com.liferay.exportimport.kernel.lar.ExportImportPathUtil;
 import com.liferay.exportimport.kernel.lar.PortletDataContext;
